Move PanZoom, FreeScroll and the Base class to a new gestures package
