+++ app/src/androidTest/java/com/spundev/dynamicthemeexport/util/TransparentLongClickTest.kt
@@ -12,9 +12,9 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onRoot
 import androidx.compose.ui.test.performTouchInput
 import androidx.compose.ui.unit.dp
-import com.spundev.dynamicthemeexport.util.panZoom.PanZoomState
-import com.spundev.dynamicthemeexport.util.panZoom.panZoom
-import com.spundev.dynamicthemeexport.util.panZoom.rememberPanZoomState
+import com.spundev.dynamicthemeexport.util.gestures.panZoom.PanZoomState
+import com.spundev.dynamicthemeexport.util.gestures.panZoom.panZoom
+import com.spundev.dynamicthemeexport.util.gestures.panZoom.rememberPanZoomState
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Rule
