+++ app/src/main/java/com/spundev/dynamicthemeexport/ui/preview/PreviewScreen.kt
@@ -50,8 +50,8 @@ import com.spundev.dynamicthemeexport.ui.preview.components.ColorBlockWithFixedA
 import com.spundev.dynamicthemeexport.ui.preview.components.DefaultColorBlockStyle
 import com.spundev.dynamicthemeexport.ui.preview.components.ForceSmallColorBlockStyle
 import com.spundev.dynamicthemeexport.ui.theme.DynamicExportTheme
-import com.spundev.dynamicthemeexport.util.freeScroll.freeScroll
-import com.spundev.dynamicthemeexport.util.freeScroll.rememberFreeScrollState
+import com.spundev.dynamicthemeexport.util.gestures.freeScroll.freeScroll
+import com.spundev.dynamicthemeexport.util.gestures.freeScroll.rememberFreeScrollState
 import kotlinx.coroutines.launch
 
 @Deprecated(
